API Reference

ADAS

Missing docstring.

Missing docstring for ADASdata. Check Documenter's build log for details.

Missing docstring.

Missing docstring for build_ADAS_database. Check Documenter's build log for details.

Missing docstring.

Missing docstring for meshgrid. Check Documenter's build log for details.

ADAS.retrieve_ADAS_data — Function

alias for retrieveelementdata

ADAS.set_adas_parsed_data_directory! — Function

set_adas_parsed_data_directory!(dir::String=joinpath(@__DIR__, "../"))

Set the directory where parsed ADAS data files are stored.

This function updates the data_paths dictionary with the specified directory for both adf11 and adf15 data types. If the specified directory does not exist, it creates the directory.

Arguments

dir::String: The base directory where the parsed data directories for adf11 and adf15 are located. Defaults to the parent directory of the current file.

ADAS.set_adas_raw_data_directory! — Function

set_adas_raw_data_directory!(dir=joinpath(@__DIR__, "../"))

Set the root directory for ADAS raw data files (adf11 and adf15).

The function iterates through the ADAS data types (:adf11 and :adf15), and sets the corresponding paths in the data_paths[:raw_data] dictionary. If the specified directory does not exist, it creates the directory.

Arguments

dir::String: The root directory where the "data" subdirectory containing "adf11" and "adf15" subdirectories are located. Defaults to the parent directory of the current file.

Missing docstring.

Missing docstring for show_ADAS_data. Check Documenter's build log for details.

Missing docstring.

Missing docstring for show_ADAS_files. Check Documenter's build log for details.

Missing docstring.

Missing docstring for show_adf11_types. Check Documenter's build log for details.